QSC: use hash instead of actual messages
